fix(actionParser): no action text return null not error
